renamed lib/blue-eyes to lib/blue_eyes to conform to convention
