+++ .rubocop.yml
@@ -1,11 +1,20 @@
 inherit_from:
   - .rubocop_todo.yml
 
-AllCops:
-  NewCops: enable
-  TargetRubyVersion: 3.0
-
 plugins:
   - rubocop-minitest
   - rubocop-performance
   - rubocop-rake
+
+AllCops:
+  NewCops: enable
+  TargetRubyVersion: 3.0
+    - 'app/db/**/*' # IGNORE: Generated by Rake db:migrate
+
+Metrics/ClassLength:
+  Exclude:
+    - 'lib/blue_eyes/generators/app/app_generator.rb' # IGNORE: Number of methods to generate app exceeds limit
+
+Style/Documentation:
+  Exclude:
+    - 'test/**/*'
