+++ .rubocop_todo.yml
@@ -1,44 +1,15 @@
-# This configuration was generated by
-# `rubocop --auto-gen-config`
-# on 2025-11-28 14:09:29 UTC using RuboCop version 1.81.7.
-# The point is for the user to remove these configuration records
-# one by one as the offenses are removed from the code base.
-# Note that changes in the inspected code, or installation of new
-# versions of RuboCop, may require this file to be generated again.
+# remaining linting tasks
 
-# Ignore offenses in generated app directory
+# Ignore (aka Exclude) offenses in generated app folder
+# NOTE: Include whenever linting while coding
 AllCops:
   Exclude:
-    - 'app/**/*' # NOTE: Include whenever linting while coding
-    - 'app/db/**/*' # IGNORE: Generated by Rake db:migrate
+    - 'app/**/*'
 
 Metrics/AbcSize:
   Exclude:
-    - 'lib/blue-eyes/generators/model/model_generator.rb' # TODO: Fix setup method
+    - 'lib/blue_eyes/generators/model/model_generator.rb' # TODO: Fix setup method
 
-Metrics/ClassLength:
-  Exclude:
-    - 'lib/blue-eyes/generators/app/app_generator.rb'
-
-# Offense count: 2
-# Configuration parameters: ExpectMatchingDefinition, CheckDefinitionPathHierarchy, CheckDefinitionPathHierarchyRoots, Regex, IgnoreExecutableScripts, AllowedAcronyms.
-# CheckDefinitionPathHierarchyRoots: lib, spec, test, src
-# AllowedAcronyms: CLI, DSL, ACL, API, ASCII, CPU, CSS, DNS, EOF, GUID, HTML, HTTP, HTTPS, ID, IP, JSON, LHS, QPS, RAM, RHS, RPC, SLA, SMTP, SQL, SSH, TCP, TLS, TTL, UDP, UI, UID, UUID, URI, URL, UTF8, VM, XML, XMPP, XSRF, XSS
-Naming/FileName:
-  Exclude:
-    - 'lib/blue-eyes.rb' # TODO: Module file name should be underscore
-    - 'test/blue-eyes_test.rb' # TODO: Fix when lib/blue-eyes.rb renamed to lib/blue_eyes.rb
-
-# Offense count: 1
-# Configuration parameters: Mode, AllowedMethods, AllowedPatterns, AllowBangMethods, WaywardPredicates.
-# AllowedMethods: call
-# WaywardPredicates: nonzero?
 Naming/PredicateMethod:
   Exclude:
-    - 'lib/blue-eyes/cli.rb' # Can't fix. Required by Thor
-
-# Offense count: 1
-# Configuration parameters: AllowedConstants.
-Style/Documentation:
-  Exclude:
-    - 'test/**/*'
+    - 'lib/blue_eyes/cli.rb' # Can't fix. Required by Thor
