+++ bin/blue-eyes
@@ -3,6 +3,6 @@
 
 $LOAD_PATH.unshift(File.expand_path(File.join(File.dirname(__FILE__), '..', 'lib')))
 
-require 'blue-eyes'
+require 'blue_eyes'
 
 BlueEyes::CLI.start
