+++ blue-eyes.gemspec
@@ -3,7 +3,7 @@
 lib = File.expand_path('lib', __dir__)
 $LOAD_PATH.unshift lib unless $LOAD_PATH.include?(lib)
 
-require 'blue-eyes/version'
+require 'blue_eyes/version'
 
 Gem::Specification.new do |spec|
   spec.name        = 'blue-eyes'
