+++ lib/blue_eyes.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'blue_eyes/extensions/string'
+require 'blue_eyes/generators/app/app_generator'
+require 'blue_eyes/generators/controller/controller_generator'
+require 'blue_eyes/generators/model/model_generator'
+require 'blue_eyes/generators/scaffold/scaffold_generator'
+require 'blue_eyes/cli'
+require 'blue_eyes/version'
+
+# BlueEyes is a Rails-style generator for Sinatra
+module BlueEyes
+end
