+++ lib/blue_eyes/cli.rb
