+++ lib/blue_eyes/generators/app/templates/app/controllers/application_controller.rb
