+++ lib/blue_eyes/generators/app/templates/app/views/layout.erb.tt
