+++ lib/blue_eyes/generators/app/templates/app/views/welcome.erb
