+++ lib/blue_eyes/generators/app/templates/config/db.yml
