+++ lib/blue_eyes/generators/app/templates/config/environment.rb
