+++ lib/blue_eyes/generators/app/templates/config/redis.yml
