+++ lib/blue_eyes/generators/app/templates/public/images/blue-eyes-small.png
