+++ lib/blue_eyes/generators/app/templates/spec/application_controller_spec.rb
