+++ lib/blue_eyes/generators/controller/controller_generator.rb
