+++ lib/blue_eyes/generators/controller/templates/views/edit.html.erb
