+++ lib/blue_eyes/generators/controller/templates/views/index.html.erb
