+++ lib/blue_eyes/generators/controller/templates/views/new.html.erb
