+++ lib/blue_eyes/generators/model/migration.rb.erb
