+++ lib/blue_eyes/generators/model/model.rb.erb
