+++ lib/blue_eyes/generators/model/model_factory.rb.erb
