+++ lib/blue_eyes/generators/model/model_generator.rb
