+++ lib/blue_eyes/generators/model/model_spec.rb.erb
