+++ lib/blue_eyes/generators/scaffold/scaffold_generator.rb
