+++ test/blue_eyes/generators/model/model_generator_test.rb
@@ -9,7 +9,7 @@ class ModelGeneratorTest < Minitest::Test
 
   def test_source_root
     root_path = File.expand_path('../../../..', __dir__)
-    absolute_template_path = [root_path, '/lib/blue-eyes/generators/model'].join
+    absolute_template_path = [root_path, '/lib/blue_eyes/generators/model'].join
 
     assert_equal subject.source_root, absolute_template_path
   end
