+++ test/blue_eyes/version_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative '../test_helper'
-# require 'blue-eyes/version'
+# require 'blue_eyes/version'
 
 class VersionTest < Minitest::Test
   def test_major_version
