+++ test/test_helper.rb
@@ -3,21 +3,4 @@
 require 'minitest/autorun'
 require 'fakefs/safe'
 require 'simplecov'
-# require_relative "../lib/blue-eyes"
-require 'blue-eyes'
-
-# require_relative "../spec_helper"
-# require_relative "../lib/blue-eyes/"
-# require_all '../lib/blue-eyes'
-
-# describe BlueEyes::VERSION do
-#   subject { BlueEyes::Generators::AppGenerator }
-
-#   describe "source_root" do
-#     it "should return with a full path" do
-#       pwd           = File.dirname(__FILE__)
-#       template_path = pwd.sub 'spec/blue-eyes', 'lib/blue-eyes/generators/app/templates'
-#       subject.source_root.must_equal template_path
-#     end
-#   end
-# end
+require 'blue_eyes'
